refactoring epc models

+++ epc-modelling/src/main/java/at/uibk/epc/model/Measure.java
@@ -2,7 +2,7 @@ package at.uibk.epc.model;
 
 public class Measure {
 
-	private long value;
+	private double value;
 	
 	private MeasuringUnit unit;
 	
@@ -10,7 +10,7 @@ public class Measure {
 		//needed by MongoDB
 	}
 
-	public Measure(long value, MeasuringUnit unit) {
+	public Measure(double value, MeasuringUnit unit) {
 		this.value = value;
 		this.unit = unit;
 	}
@@ -23,11 +23,11 @@ public class Measure {
 		this.unit = unit;
 	}
 	
-	public long getValue() {
+	public double getValue() {
 		return value;
 	}
 	
-	public void setValue(long value) {
+	public void setValue(double value) {
 		this.value = value;
 	}
